fix compile warning

+++ mod_sndmem.cpp
@@ -217,8 +217,8 @@ static switch_status_t sndfile_file_open(switch_file_handle_t *handle, const cha
 	map = static_cast<format_map *>(switch_core_hash_find(globals.format_hash, ext));
 
 	if (mode & SFM_WRITE) {
-		context->sfinfo.channels = handle->channels;
-		context->sfinfo.samplerate = handle->samplerate;
+		context->sfinfo.channels = (int)handle->channels;
+		context->sfinfo.samplerate = (int)handle->samplerate;
 		if (handle->samplerate == 8000 || handle->samplerate == 16000 ||
 			handle->samplerate == 24000 || handle->samplerate == 32000 || handle->samplerate == 48000 ||
 			handle->samplerate == 11025 || handle->samplerate == 22050 || handle->samplerate == 44100) {
